simplification in tokenGated files; fixes on pin&promote for erc20, erc721
