+++ src/features/room/sagas/pin.saga.ts
@@ -2,8 +2,8 @@ import RoomNotFoundError from '$/errors/RoomNotFoundError'
 import { PreferencesAction } from '$/features/preferences'
 import { RoomAction } from '$/features/room'
 import { IRoom, RoomId } from '$/features/room/types'
-import { isTokenGatedRoom } from '$/features/tokenGatedRooms/utils/isTokenGatedRoom'
-import joinTokenGatedRoom from '$/features/tokenGatedRooms/utils/joinTokenGatedRoom'
+import { TokenGatedRoomAction } from '$/features/tokenGatedRooms'
+import { isTokenGatedRoom } from '$/features/tokenGatedRooms/utils'
 import { Address, EnhancedStream } from '$/types'
 import db from '$/utils/db'
 import getStream from '$/utils/getStream'
@@ -13,7 +13,7 @@ import takeEveryUnique from '$/utils/takeEveryUnique'
 import { error } from '$/utils/toaster'
 import { Provider } from '@web3-react/types'
 import { toast } from 'react-toastify'
-import { call, put } from 'redux-saga/effects'
+import { call, put, putResolve } from 'redux-saga/effects'
 import StreamrClient from 'streamr-client'
 
 interface PinRemoteOptions {
@@ -33,22 +33,19 @@ function* pinRemote(
         throw new RoomNotFoundError(roomId)
     }
 
-    const metadata = stream.extensions['thechat.eth']
-
-    const { createdAt, createdBy, tokenAddress } = stream.extensions['thechat.eth']
+    const { createdAt, createdBy } = stream.extensions['thechat.eth']
 
     const isTokenGated = isTokenGatedRoom(stream)
 
-    if (isTokenGated && tokenAddress && metadata.tokenType) {
-        yield joinTokenGatedRoom(
-            roomId,
-            owner,
-            tokenAddress,
-            provider,
-            delegatedAccount,
-            metadata.tokenType,
-            metadata.tokenId!,
-            streamrClient
+    if (isTokenGated) {
+        yield putResolve(
+            TokenGatedRoomAction.join({
+                roomId,
+                owner,
+                provider,
+                delegatedAccount,
+                streamrClient,
+            })
         )
     }
     const room: undefined | IRoom = yield db.rooms.where({ id: stream.id, owner }).first()
