+++ src/features/tokenGatedRooms/sagas/registerPolicy.saga.ts
@@ -3,7 +3,7 @@ import { Contract, StreamPermission } from 'streamr-client'
 import { BigNumber } from 'ethers'
 import { Address } from '$/types'
 import { TokenGatedRoomAction } from '..'
-import { getJoinPolicyFactory } from '$/features/tokenGatedRooms/utils/getJoinPolicyFactory'
+import { getJoinPolicyFactory } from '$/features/tokenGatedRooms/utils'
 import { toast } from 'react-toastify'
 import handleError from '$/utils/handleError'
 import { error } from '$/utils/toaster'
