--- src/features/tokenGatedRooms/utils/isTokenGatedRoom.ts
@@ -1,5 +1,0 @@
-import { EnhancedStream } from '$/types'
-
-export function isTokenGatedRoom(stream: EnhancedStream): boolean {
-    return !!stream.extensions['thechat.eth'].tokenAddress
-}
